+++ src/main/java/de/joh/dragonmagicandrelics/spells/components/ComponentAlternativeRecall.java
@@ -11,21 +11,13 @@ import com.mna.api.spells.parts.SpellEffect;
 import com.mna.api.spells.targeting.SpellContext;
 import com.mna.api.spells.targeting.SpellSource;
 import com.mna.api.spells.targeting.SpellTarget;
-import com.mna.items.ItemInit;
-import com.mna.items.runes.MarkBookItem;
 import com.mna.tools.TeleportHelper;
 import de.joh.dragonmagicandrelics.utils.MarkSave;
-import de.joh.dragonmagicandrelics.capabilities.dragonmagic.PlayerDragonMagicProvider;
 import net.minecraft.Util;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.TranslatableComponent;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.level.Level;
 import net.minecraft.world.phys.Vec3;
-import javax.annotation.Nullable;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicReference;
 
 public class ComponentAlternativeRecall extends SpellEffect {
     public ComponentAlternativeRecall(ResourceLocation registryName, ResourceLocation guiIcon) {
@@ -42,11 +34,11 @@ public class ComponentAlternativeRecall extends SpellEffect {
             if (markSave != null) {
                 BlockPos pos = markSave.getPosition();
                 double dist = target.getEntity().blockPosition().distSqr(pos);
-                double maxDist = (double)(modificationData.getValue(Attribute.MAGNITUDE) * 1000.0F);
+                double maxDist = modificationData.getValue(Attribute.MAGNITUDE) * 1000.0F;
                 if (!(dist > maxDist * maxDist)) {
                     int magnitude = (int)modificationData.getValue(Attribute.MAGNITUDE);
                     if (this.magnitudeHealthCheck(source, target, magnitude, 20)) {
-                        TeleportHelper.teleportEntity(target.getEntity(), context.getWorld().dimension(), new Vec3((double)pos.getX() + 0.5, (double)(pos.getY() + 1), (double)pos.getZ() + 0.5));
+                        TeleportHelper.teleportEntity(target.getEntity(), context.getWorld().dimension(), new Vec3((double)pos.getX() + 0.5, pos.getY() + 1, (double)pos.getZ() + 0.5));
                         return ComponentApplicationResult.SUCCESS;
                     }
